% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breathe.R
\name{breathe}
\alias{breathe}
\title{Take Time Out To Breathe}
\usage{
breathe(times = 3L, seconds = 5)
}
\arguments{
\item{times}{Numeric. Number of times to breathe.}

\item{seconds}{Numeric. Number of seconds to take when inhaling and exhaling.}
}
\value{
Message printed to the console.
}
\description{
Take Time Out To Breathe
}
\examples{
\donttest{breathe()}
}
