% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-rd.R
\name{rd_email}
\alias{rd_email}
\title{Link to Email}
\usage{
rd_email()
}
\description{
Surround selected text with Rd markup for an email link.
}
\details{
\itemize{
\item User-selected text: \code{address}
\item Result of function: \verb{\email{address}}
\item Example output with no text selection: \verb{\email{}}
}
}
\references{
\href{https://cran.r-project.org/doc/manuals/R-exts.html#Writing-R-documentation-files}{The \emph{Writing R Extensions} manual.}
}
