% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-rd.R
\name{rd_fn_other_pkg}
\alias{rd_fn_other_pkg}
\title{Link To Function (Another Package)}
\usage{
rd_fn_other_pkg()
}
\description{
Surround selected text (in the form 'package::function') with Rd markup that
links to a function in another package.
}
\details{
Example output: \verb{\code{\link[package]{function}}}
}
\references{
\href{https://cran.r-project.org/doc/manuals/R-exts.html#Writing-R-documentation-files}{The \emph{Writing R Extensions} manual.}
}
