% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-rd.R
\name{rd_link_fn_other_pkg_name}
\alias{rd_link_fn_other_pkg_name}
\title{Link To Function (Another Package, Show Name)}
\usage{
rd_link_fn_other_pkg_name()
}
\description{
Surround selected text (in the form 'package::function') with Rd markup that
links to a function in another package and shows that package's name.
}
\details{
Example output: \verb{\code{\link[package:function]}}
}
\references{
\href{https://cran.r-project.org/doc/manuals/R-exts.html#Writing-R-documentation-files}{The \emph{Writing R Extensions} manual.}
}
