% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-rd.R
\name{rd_link_s4_class}
\alias{rd_link_s4_class}
\title{Link To S4 Class}
\usage{
rd_link_s4_class()
}
\description{
Surround selected text with Rd markup to link to an S4 class.
}
\details{
Example output: \verb{\linkS4class{class}}
}
\references{
\href{https://cran.r-project.org/doc/manuals/R-exts.html#Writing-R-documentation-files}{The \emph{Writing R Extensions} manual.}
}
