% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-rd.R
\name{rd_list_definition}
\alias{rd_list_definition}
\title{Insert List (Named)}
\usage{
rd_list_definition()
}
\description{
Insert an Rd markup skeleton for a definition (named) list.
}
\details{
Example output:\preformatted{\\describe\{
  \\item\{One\}\{First item\}
  \\item\{Two\}\{Second item\}
\}

}
}
\references{
\href{https://cran.r-project.org/doc/manuals/R-exts.html#Writing-R-documentation-files}{The \emph{Writing R Extensions} manual.}
}
