% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-rd.R
\name{rd_list_ordered}
\alias{rd_list_ordered}
\title{Insert List (Ordered)}
\usage{
rd_list_ordered()
}
\description{
Insert an Rd markup skeleton for an ordered (numbered) list.
}
\details{
Example output:\preformatted{\\enumerate\{
  \\item First item
  \\item Second item
\}

}
}
\references{
\href{https://cran.r-project.org/doc/manuals/R-exts.html#Writing-R-documentation-files}{The \emph{Writing R Extensions} manual.}
}
