% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-rd.R
\name{rd_list_unordered}
\alias{rd_list_unordered}
\title{Insert List (Bulleted)}
\usage{
rd_list_unordered()
}
\description{
Insert an Rd markup skeleton for an unordered (bulleted) list.
}
\details{
Example output:\preformatted{\\itemize\{
  \\item First item
  \\item Second item
\}

}
}
\references{
\href{https://cran.r-project.org/doc/manuals/R-exts.html#Writing-R-documentation-files}{The \emph{Writing R Extensions} manual.}
}
