% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-rd.R
\name{rd_package}
\alias{rd_package}
\title{Format Package}
\usage{
rd_package()
}
\description{
Surround selected text with Rd markup for package names.
}
\details{
\itemize{
\item User-selected text: \code{package}
\item Result of function: \verb{\pkg{package}}
\item Example output with no text selection: \verb{\pkg{}}
}
}
\references{
\href{https://cran.r-project.org/doc/manuals/R-exts.html#Writing-R-documentation-files}{The \emph{Writing R Extensions} manual.}
}
