% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snorkel-package.R
\docType{package}
\name{snorkel-package}
\alias{snorkel}
\alias{snorkel-package}
\title{snorkel: RStudio Addins To Help Format 'roxygen2'}
\description{
Insert, or replace text with, formatting code for 'roxygen2' function documentation in a package, like making text bold, inserting a URL, or linking to functions in other packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/snorkel}
  \item Report bugs at \url{https://github.com/matt-dray/snorkel/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
