% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-rd.R
\name{rd_bold}
\alias{rd_bold}
\title{Format Bold}
\usage{
rd_bold()
}
\description{
Surround selected text with Rd markup for bold.
}
\details{
\itemize{
\item Selected text: \code{text}
\item Result of function: \verb{\\strong\{text\}}
\item Example output with no text selection: \verb{\\strong\{\}}
}
}
\references{
\href{https://cran.r-project.org/doc/manuals/R-exts.html#Writing-R-documentation-files}{The \emph{Writing R Extensions} manual.}
}
