% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-rd.R
\name{rd_code}
\alias{rd_code}
\title{Format Code}
\usage{
rd_code()
}
\description{
Surround selected text with Rd markup for inline code.
}
\details{
Example output: \verb{\code{text}}
}
\references{
\href{https://cran.r-project.org/doc/manuals/R-exts.html#Writing-R-documentation-files}{The \emph{Writing R Extensions} manual.}
}
