% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-rd.R
\name{rd_fn_this_pkg}
\alias{rd_fn_this_pkg}
\title{Link to Function (This Package)}
\usage{
rd_fn_this_pkg()
}
\description{
Surround selected text with Rd markup that links to a function in the
current package.
}
\details{
Example output: \verb{\code{\link{text}}}
}
\references{
\href{https://cran.r-project.org/doc/manuals/R-exts.html#Writing-R-documentation-files}{The \emph{Writing R Extensions} manual.}
}
