% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-rd.R
\name{rd_italic}
\alias{rd_italic}
\title{Format Italic}
\usage{
rd_italic()
}
\description{
Surround selected text with Rd markup for italic.
}
\details{
Example output: \verb{\\emph\{text\}}
}
\references{
\href{https://cran.r-project.org/doc/manuals/R-exts.html#Writing-R-documentation-files}{The \emph{Writing R Extensions} manual.}
}
