% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-rd.R
\name{rd_url}
\alias{rd_url}
\title{Link To Web}
\usage{
rd_url()
}
\description{
Surround selected URL text with Rd markup for a link to the web.
}
\details{
\itemize{
\item User-selected text: \code{url}
\item Result of function: \verb{\url{url}}
\item Example output with no text selection: \verb{\url{}}
}
}
\references{
\href{https://cran.r-project.org/doc/manuals/R-exts.html#Writing-R-documentation-files}{The \emph{Writing R Extensions} manual.}
}
