% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-rd.R
\name{rd_url_named}
\alias{rd_url_named}
\title{Link To Web (Named)}
\usage{
rd_url_named()
}
\description{
Surround selected URL text with Rd markup for a link to the web and show a
text label.
}
\details{
\itemize{
\item User-selected text: \code{url}
\item Result of function: \verb{\\href\{url\}\{Link text\}}
\item Example output with no text selection: \verb{\\href\{\}\{Link text\}}
}
}
\references{
\href{https://cran.r-project.org/doc/manuals/R-exts.html#Writing-R-documentation-files}{The \emph{Writing R Extensions} manual.}
}
