% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snorkel-package.R
\docType{package}
\name{snorkel-package}
\alias{snorkel}
\alias{snorkel-package}
\title{snorkel: RStudio Addins To Format 'roxygen2'}
\description{
Insert formatting code for 'roxygen2' functtion documentation.
}
\author{
\strong{Maintainer}: Matt Dray \email{medray@gmail.com}

}
\keyword{internal}
